Move CBOR encode/decode to Cbor class
